remove auth actions and move content to lib/auth/action.ts
